Move app code to separate file for use with supertest
